[scripts] add agent state address auto-update to buildnfuzz script
With the approach of passing agent state accross modules via a globally visible, fixed memory address, it is necessary to update the hardcoded memory address in kAFL agent library to match the real address from a fuzzing session. This can be done by running the session, observing the real address of the agent state struct & then updateing the address accordingly in the kAFL agent lib. This commit adds this process to the build-n-fuzz script so it is performed in an automated way.